Move Rendering into concern
