+++ spec/node/render_spec.rb
